fix : 드랍 아이템 구조화하여 설계
- ExpGem, Coin 등 드랍 아이템 구조화하여 설계
