+++ Heroes_vs_Hordes/Assets/Scripts/Objects/Items/DropItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2d7aaa92e0ad008429fc7183d69bc889
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
