+++ Heroes_vs_Hordes/Assets/Scripts/Objects/Items/Gold.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Gold : DropItem
+{
+    protected override void _DeliverEffect(Hero targetHero, bool getEffect)
+    {
+
+    }
+}
