+++ Heroes_vs_Hordes/Assets/Scripts/Objects/Items/Gold.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 901d2a94b7d28f849ab967ed0ed10844
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
